:recycle: :: DetailEventNavigation을 한 파일에 담고 코드를 리팩터링합니다.
